ruby: Ensure instructions are complete for ruby-1-a
There were a few things that were unclear.

- the SKU should be a Symbol rather than a String. This was not
specified, and the example showed a constant, which doesn't tell the
research subject what the type is.
- the price list shows a combination of integers and floats. The
instructions state that the total amount should have 2 decimal places,
but the example shows a return value of an integer with no decimal
places.
- the example shows the item names in lowercase, whereas the actual
table shows the names capitalized.
- the README doesn't specify that the `item_list` requires the names
to be unique.

I also updated the tests to assert against floats consistently, even
though Ruby considers 1 == 1.0.